Fixed bugs with UI interaction
IMPORTANT: I commented out the "Don't Destroy on Load" code and instead put a Game Manager object in each scene. This was so that the UI buttons would transition from scene to scene properly. This change MIGHT cause trouble if we are trying to save values from scene to scene (e.g. health, score, etc.)