Modified UI so that the Music and SFX sliders retain the same value even when switching from scene to scene.
I created a new C# script that will save our volume settings from scene to scene. Like with the GameManager, I put a SoundManager object in each scene. I also put dummy inactive sliders in each scene so that I don't get errors (and so we can see the value on the slider in the inspector). There's probably a better way to do this, but I don't think we have time to figure that out.

I also moved all of the scripts to the "Scripts" folder as some of them were outside that folder.